+++ src/main/java/com/project/expensemanage/notification/recommendation/discord/body/embedObject/DiscordEmbedThumbnail.java
@@ -1,4 +1,4 @@
-package com.project.expensemanage.recommendation.discord.body.embedObject;
+package com.project.expensemanage.notification.recommendation.discord.body.embedObject;
 
 import lombok.Builder;
