+++ src/main/java/com/project/expensemanage/notification/recommendation/dto/RecommendationExpenditure.java
@@ -1,4 +1,4 @@
-package com.project.expensemanage.recommendation.dto;
+package com.project.expensemanage.notification.recommendation.dto;
 
 
 public record RecommendationExpenditure (Long categoryId, String categoryName,Long totalExpenditure, Long budget){}
