+++ src/main/java/com/project/expensemanage/notification/recommendation/dto/RecommendationExpenditureAllUser.java
@@ -1,4 +1,4 @@
-package com.project.expensemanage.recommendation.dto;
+package com.project.expensemanage.notification.recommendation.dto;
 
 
 public record RecommendationExpenditureAllUser(Long categoryId, Long userId,String categoryName, Long totalExpenditure, Long budget){}
