+++ src/main/java/com/project/expensemanage/notification/recommendation/event/event/DailyRecommendationExpenditureEvent.java
@@ -1,7 +1,7 @@
-package com.project.expensemanage.recommendation.event.event;
+package com.project.expensemanage.notification.recommendation.event.event;
 
 import com.project.expensemanage.domain.user.entity.User;
-import com.project.expensemanage.recommendation.dto.RecommendationExpenditure;
+import com.project.expensemanage.notification.recommendation.dto.RecommendationExpenditure;
 import java.util.List;
 
 public record DailyRecommendationExpenditureEvent(List<RecommendationExpenditure> recommendationList, User user) {
