+++ src/main/java/com/project/expensemanage/notification/recommendation/event/listener/ExpenditureRecommendationEventListener.java
@@ -1,6 +1,6 @@
-package com.project.expensemanage.recommendation.event.listener;
+package com.project.expensemanage.notification.recommendation.event.listener;
 
-import com.project.expensemanage.recommendation.event.event.DailyRecommendationExpenditureEvent;
+import com.project.expensemanage.notification.recommendation.event.event.DailyRecommendationExpenditureEvent;
 import org.springframework.context.event.EventListener;
 
 public interface ExpenditureRecommendationEventListener {
