--- src/main/java/com/project/expensemanage/recommendation/discord/body/embedObject/DiscordEmbedImage.java
@@ -1,3 +1,0 @@
-package com.project.expensemanage.recommendation.discord.body.embedObject;
-
-public record DiscordEmbedImage() {}
