+++ src/test/java/com/project/expensemanage/recommendation/repository/RecommendationRepositoryTest.java
@@ -3,9 +3,10 @@ package com.project.expensemanage.recommendation.repository;
 import com.project.expensemanage.commone.config.QueryDslConfig;
 import com.project.expensemanage.domain.budget.repository.BudgetRepository;
 import com.project.expensemanage.domain.expenditure.repoistory.ExpenditureRepository;
+import com.project.expensemanage.notification.recommendation.repository.RecommendationRepository;
 import com.project.expensemanage.recommendation.config.TestSQLUtils;
-import com.project.expensemanage.recommendation.dto.RecommendationExpenditure;
-import com.project.expensemanage.recommendation.dto.RecommendationExpenditureAllUser;
+import com.project.expensemanage.notification.recommendation.dto.RecommendationExpenditure;
+import com.project.expensemanage.notification.recommendation.dto.RecommendationExpenditureAllUser;
 import java.time.LocalDate;
 import java.util.List;
 import org.assertj.core.api.Assertions;
